backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
mangum>=0.17.0
boto3>=1.34.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1
